Move TransactionConditional to Bundle file
