+++ crates/op-rbuilder/src/primitives/bundle.rs
@@ -1,5 +1,8 @@
 use alloy_primitives::Bytes;
+use alloy_rpc_types_eth::erc4337::TransactionConditional;
 use serde::{Deserialize, Serialize};
+
+pub const MAX_BLOCK_RANGE_BLOCKS: u64 = 10;
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct Bundle {
@@ -10,4 +13,14 @@ pub struct Bundle {
     pub block_number_max: Option<u64>,
 }
 
-pub const MAX_BLOCK_RANGE_BLOCKS: u64 = 10;
+impl Bundle {
+    pub fn conditional(&self) -> TransactionConditional {
+        TransactionConditional {
+            block_number_min: None,
+            block_number_max: self.block_number_max,
+            known_accounts: Default::default(),
+            timestamp_max: None,
+            timestamp_min: None,
+        }
+    }
+}
